Undo removing CPU kernel to impl.h
